+++ poc/classes/AuxSTConjecture.py
@@ -1,13 +1,8 @@
-from poc.classes.AuxSTBlockWithSignature import AuxSTBlockWithSignature
+from poc.classes.AuxSTTheoremLikeStatementOrConjecture import AuxSTTheoremLikeStatementOrConjecture
 from poc.classes.AuxSymbolTable import AuxSymbolTable
 
 
-class AuxSTConjecture(AuxSTBlockWithSignature):
-
+class AuxSTConjecture(AuxSTTheoremLikeStatementOrConjecture):
     def __init__(self, i):
-        super().__init__(AuxSymbolTable.block_conj, i)
-        self.zfrom = i.corrected_position('TheoremLikeStatementOrConjectureHeader')
-        self.zto = i.last_positions_by_rule['TheoremLikeStatementOrConjecture'].pos_to_str()
-
-    def evaluate(self, sem):
-        raise NotImplementedError()
+        super().__init__(AuxSymbolTable.block_conj, i, i.corrected_position('TheoremLikeStatementOrConjectureHeader'),
+                         i.last_positions_by_rule['TheoremLikeStatementOrConjecture'].pos_to_str())