+++ poc/classes/AuxSTLemma.py
@@ -1,13 +1,9 @@
-from poc.classes.AuxSTBlockWithSignature import AuxSTBlockWithSignature
+from poc.classes.AuxSTTheoremLikeStatementOrConjecture import AuxSTTheoremLikeStatementOrConjecture
 from poc.classes.AuxSymbolTable import AuxSymbolTable
 
 
-class AuxSTLemma(AuxSTBlockWithSignature):
+class AuxSTLemma(AuxSTTheoremLikeStatementOrConjecture):
 
     def __init__(self, i):
-        super().__init__(AuxSymbolTable.block_lem, i)
-        self.zfrom = i.corrected_position('TheoremLikeStatementOrConjectureHeader')
-        self.zto = i.last_positions_by_rule['TheoremLikeStatementOrConjecture'].pos_to_str()
-
-    def evaluate(self, sem):
-        raise NotImplementedError()
+        super().__init__(AuxSymbolTable.block_lem, i, i.corrected_position('TheoremLikeStatementOrConjectureHeader'),
+                         i.last_positions_by_rule['TheoremLikeStatementOrConjecture'].pos_to_str())