+++ poc/classes/AuxSTPredicateWithArgs.py
@@ -1,10 +1,10 @@
-from poc.classes.AuxST import AuxSTBlock
+from poc.classes.AuxSTBlock import AuxSTBlock
 from poc.classes.AuxSymbolTable import AuxSymbolTable
 from poc.classes.AuxSTVariable import AuxSTVariable
 from poc.classes.AuxSTSelf import AuxSTSelf
 from poc.classes.AuxSTStatement import AuxSTStatement
-from poc.classes.AuxInbuiltTypes import InbuiltUndefined
-from fplerror import FplIdentifierNotDeclared, FplWrongArguments
+from poc.classes.AuxInbuiltTypes import InbuiltUndefined, InbuiltPredicate, EvaluatedPredicate
+from fplerror import FplIdentifierNotDeclared, FplWrongArguments, FplPredicateRecursion, FplVariableNotInitialized
 from poc.classes.AuxEvaluation import EvaluateParams
 
 
@@ -22,9 +22,6 @@ class AuxSTPredicateWithArgs(AuxSTBlock):
         new_predicate._value = self._value
         return new_predicate
 
-    def get_type_signature(self):
-        return self.reference.get_type_signature()
-
     def evaluate(self, sem):
         """
         Checks if
@@ -36,22 +33,39 @@ class AuxSTPredicateWithArgs(AuxSTBlock):
         :param sem:
         :return:
         """
-        self.reference = None
-        self._calculate_argument_types(sem)
-        self._check_if_referenced_id_exists(sem)
-        # at this stage, there is exactly one override determined (even if is an InbuiltUndefined instance)
-        if isinstance(self.reference, InbuiltUndefined):
+        arg_list = self._calculate_arguments(sem)
+        possible_overrides = self._check_if_referenced_id_exists(sem)
+        if len(possible_overrides) == 0:
+            # if no possible overrides found
             sem.eval_stack[-1].value = InbuiltUndefined()
             return
-        # at this stage, we can call the predicate with args
-        propagated_expected_type = sem.eval_stack[-1].expected_type
-        arg_list = AuxSymbolTable.get_child_by_outline(self, AuxSymbolTable.arg_list)
-        # recursive call
-        ret = EvaluateParams.evaluate_recursion(sem, self.reference, propagated_expected_type, list(arg_list.children))
+
+        # at this stage, we have a list of possible overrides, we now try to "call" them with the given arguments
+        mismatched_overrides = list()
+        for override in possible_overrides:
+            propagated_expected_type = sem.eval_stack[-1].expected_type
+            if self._check_illegal_recursion(sem, override):
+                sem.analyzer.error_mgr.add_error(FplPredicateRecursion(override.reference, self))
+                sem.eval_stack[-1].value = InbuiltUndefined()
+                return
+            ret = EvaluateParams.evaluate_recursion(sem, override.reference, propagated_expected_type, arg_list, True)
+            if ret.argument_error is not None:
+                mismatched_overrides.append(str(ret.argument_error))
+            else:
+                # break searching for matching overrides, if one was found
+                break
+
+        if len(mismatched_overrides) >= len(possible_overrides):
+            arg_types = list()
+            for type_node in ret.arg_type_list:
+                arg_types.append(type_node.to_string2())
+            sem.analyzer.error_mgr.add_error(FplWrongArguments(arg_types, mismatched_overrides, self))
+            ret.returned_value = None
+
         if ret.returned_value is None:
             sem.eval_stack[-1].value = InbuiltUndefined()
         else:
-            EvaluateParams.propagate(sem, ret)
+            sem.eval_stack[-1].value = EvaluatedPredicate(ret.returned_value.get_repr())
         return
 
     def _check_if_referenced_id_exists(self, sem):
@@ -70,74 +84,39 @@ class AuxSTPredicateWithArgs(AuxSTBlock):
                 # the reference of the predicate_with_args is never declared
                 sem.analyzer.error_mgr.add_error(
                     FplIdentifierNotDeclared(qualified_identifier, self.path[1].file_name, self.zfrom))
-                self.reference = InbuiltUndefined()
+                return list()
             else:
-                self._match_override(sem, sem.overridden_qualified_ids.get(qualified_identifier))
+                return sem.overridden_qualified_ids.get(qualified_identifier)
         elif self.outline == AuxSymbolTable.var:
-            self._match_override([self.get_declared_type().get_type_node()])
+            return [self.get_declared_type()]
         else:
             raise NotImplementedError()
 
-    def _match_override(self, sem, expected_overrides: list):
-        matching_overrides = list()
-        for override in expected_overrides:
-            if self._arguments_match_override(sem, override.reference):
-                matching_overrides.append(override.reference)
-        if len(matching_overrides) == 0:
-            sem.analyzer.error_mgr.add_error(FplWrongArguments(self.__calc_signature_from_args(sem), self))
-            self.reference = InbuiltUndefined()
-        elif len(matching_overrides) > 1:
-            raise NotImplementedError()
-        else:
-            self.reference = matching_overrides[0]
-
-    def _arguments_match_override(self, sem, node):
-        arg_types_list = sem.eval_stack[-1].arg_types_list
-        signature_node = AuxSymbolTable.get_child_by_outline(node, AuxSymbolTable.signature)
-        if len(arg_types_list) == 0 and len(signature_node.children) == 0:
-            return True
-        for child in signature_node.children:
-            raise NotImplementedError()
-
-    def _calculate_argument_types(self, sem):
+    def _calculate_arguments(self, sem):
         # create a list of argument types to be matched with any of the available overrides
         args = AuxSymbolTable.get_child_by_outline(self, AuxSymbolTable.arg_list)
-        arg_types_list = sem.eval_stack[-1].arg_types_list
-        arg_value_list = sem.eval_stack[-1].arg_value_list
+        arg_type_list = list()
         for argument_node in args.children:
             if isinstance(argument_node, AuxSTPredicateWithArgs):
                 # recursive call of for a new argument being itself a predicate with arguments
                 argument_node.evaluate(sem)
-                arg_types_list.append(argument_node.get_type_signature())
-                arg_value_list.append(argument_node.get_value())
+                arg_type_list.append(argument_node.get_declared_type())
             elif isinstance(argument_node, AuxSTVariable):
-                arg_types_list.append(argument_node.get_declared_type().get_type_signature())
-                arg_value_list.append(argument_node.get_value())
+                if not argument_node.is_bound_or_initialized():
+                    sem.analyzer.error_mgr.add_error(FplVariableNotInitialized(argument_node))
+                arg_type_list.append(argument_node.get_declared_type())
             elif isinstance(argument_node, AuxSTSelf):
-                arg_types_list.append(argument_node.get_declared_type().get_type_signature())
-                arg_value_list.append(argument_node)
+                arg_type_list.append(argument_node.get_declared_type())
             elif isinstance(argument_node, AuxSTStatement):
-                arg_types_list.append(AuxSymbolTable.undefined)
-                arg_value_list.append(argument_node.get_value())
+                arg_type_list.append(argument_node.get_declared_type())
             else:
-                arg_types_list.append(argument_node.get_type_signature())
-                arg_value_list.append(argument_node.get_value())
+                arg_type_list.append(argument_node.get_declared_type())
+        return arg_type_list
 
-    def __calc_signature_from_args(self, sem):
-        ret = list()
-        running_type = ""
-        type_counter = 0
-        arg_types_list = sem.eval_stack[-1].arg_types_list
-        for current_type in arg_types_list:
-            if running_type != current_type:
-                if running_type != "":
-                    ret.append(str(type_counter) + ":" + running_type)
-                type_counter = 1
-                running_type = current_type
-            else:
-                type_counter += 1
-        if type_counter == 1 and len(ret) == 0:
-            ret.append(str(type_counter) + ":" + running_type)
-        elif type_counter > 0 and len(ret) == 0:
-            ret.append(str(type_counter) + ":" + running_type)
-        return self.reference.get_qualified_id() + "[" + ",".join(ret) + "]"
+    def _check_illegal_recursion(self, sem, override):
+        if not isinstance(sem.eval_stack[-1].expected_type, (InbuiltPredicate, EvaluatedPredicate)):
+            return False
+        else:
+            for eval_params in sem.eval_stack:
+                if eval_params.node == override.reference:
+                    return True