+++ poc/classes/AuxSTProperties.py
@@ -1,3 +1,5 @@
+from poc.classes.AuxEvaluation import EvaluateParams
+from poc.classes.AuxInbuiltTypes import InbuiltUndefined
 from poc.classes.AuxST import AuxSTOutline
 from poc.classes.AuxSymbolTable import AuxSymbolTable
 
@@ -8,5 +10,6 @@ class AuxSTProperties(AuxSTOutline):
         super().__init__(None, AuxSymbolTable.properties)
 
     def evaluate(self, sem):
-        # todo: implement
-        pass
+        for child in self.children:
+            EvaluateParams.evaluate_recursion(sem, child, InbuiltUndefined())
+        sem.eval_stack[-1].value = InbuiltUndefined()