+++ poc/tests/errors_evaluate/test_AllEvaluateTests.py
@@ -19,6 +19,10 @@ class AllEvaluateTests(UtilTestCase):
         ("test_FplAxiomNotSatisfiable_02d.fpl", "SE0240"),
         ("test_FplAxiomNotSatisfiable_02e.fpl", "SE0240"),
         ("test_FplAxiomNotSatisfiable_03.fpl", "SE0240"),
+        ("test_FplPremiseNotSatisfiable_01.fpl", "SE0245"),
+        ("test_FplPredicateRecursion_01.fpl", "SE0260"),
+        ("test_FplPredicateRecursion_02.fpl", "SE0260"),
+
     ])
     def test_errors(self, use_case, diagnose_id):
         super().semantical_analysis_detects_fpl_error(self.folder + "/" + use_case, diagnose_id)